Termbox2

Description

Bindings to the C-language termbox2 console UI library.

Termbox2 monad

data Termbox2 a

runTermbox2 :: Termbox2 a -> IO (Either Tb2Err a)

init :: Termbox2 ()

initFile :: String -> Termbox2 ()

initFd :: Integral n => n -> Termbox2 ()

initRwFd :: (Integral n, Integral o) => n -> o -> Termbox2 ()

shutdown :: Termbox2 ()

Call this when you're finished or your terminal will act funky after exit!

width :: Integral n => Termbox2 n

Width of the drawing space in characters.

height :: Integral n => Termbox2 n

Height of the drawing space in lines.

present :: Termbox2 ()

Draws the buffer to the screen.

clear :: Termbox2 ()

Clears the screen.

setClearAttrs :: Tb2ColorAttr -> Tb2ColorAttr -> Termbox2 ()

Specify the foreground and background attributes to be applied when clearing the buffer.

setCursor :: Int -> Int -> Termbox2 ()

Set the location of the cursor (upper-left character is origin).

hideCursor :: Termbox2 ()

Hide the mouse pointer.

setCell

Arguments

:: Int

-> Int

-> Int32
Unicode code point

-> Tb2ColorAttr

-> Tb2ColorAttr

-> Termbox2 ()

Draw a single cell on the screen.

setInputMode :: Tb2Input -> Termbox2 (Maybe Tb2Input)

NB: If the argument is inputCurrent then the function acts as a query and returns the current input mode.

setOutputMode :: Tb2Output -> Termbox2 (Maybe Tb2Output)

NB: If the argument is outputCurrent then the function acts as a query and returns the current output mode.

print :: Int -> Int -> Tb2ColorAttr -> Tb2ColorAttr -> String -> Termbox2 ()

Prints a string of text to the screen.

pollEvent :: Termbox2 Tb2Event

Blocks until an exception is thrown or an event is observed.

peekEvent :: Int -> Termbox2 (Maybe Tb2Event)

Blocks for the specified number of MILLISECONDS until an exception is thrown or an event is received; returns Nothing if the timeout is reached without incident or event.

Events

data Tb2Event

Incoming event from the tty.

Constructors

Tb2Event

Fields

_type :: Tb2EventType

one of TB_EVENT_* constants

_mod :: Tb2Mod

bitwise TB_MOD_* constants

_key :: Tb2Key

one of TB_KEY_* constants

_ch :: Int32

a Unicode code point

_w :: Int32

resize width

_h :: Int32

resize height

_x :: Int32

mouse x

_y :: Int32

mouse y

Constants

Errors

newtype Tb2Err

Constructors

Tb2Err CInt

errOk :: Tb2Err

errErr :: Tb2Err

errNeedMore :: Tb2Err

errInitAlready :: Tb2Err

errInitOpen :: Tb2Err

errMem :: Tb2Err

errNoEvent :: Tb2Err

errNoTerm :: Tb2Err

errNotInit :: Tb2Err

errOutOfBounds :: Tb2Err

errRead :: Tb2Err

errResizeIOCTL :: Tb2Err

errResizePipe :: Tb2Err

Enables writing text-based user interfaces with termbox2.

errResizeSigaction :: Tb2Err

errPoll :: Tb2Err

errUnsupportedTerm :: Tb2Err

errResizeWrite :: Tb2Err

errResizePoll :: Tb2Err

Allocates the Tb2Event struct pointer, runs the UI, and frees.

errResizeRead :: Tb2Err

errResizeSscanf :: Tb2Err

Must be called before anything else. The termbox2 documentation notes that handling some exceptions requires calling shutdown followed by init again, hence this is not invoked automatically by runTermbox2.

errCapCollision :: Tb2Err

Keys

newtype Tb2Key

Constructors

Tb2Key CInt

keyCtrlTilde :: Tb2Key

keyCtrl2 :: Tb2Key

keyCtrlA :: Tb2Key

keyCtrlB :: Tb2Key

keyCtrlC :: Tb2Key

keyCtrlD :: Tb2Key

keyCtrlE :: Tb2Key

keyCtrlF :: Tb2Key

keyCtrlG :: Tb2Key

keyBackspace :: Tb2Key

keyCtrlH :: Tb2Key

keyCtrlTab :: Tb2Key

keyCtrlI :: Tb2Key

keyCtrlJ :: Tb2Key

keyCtrlK :: Tb2Key

keyCtrlL :: Tb2Key

keyCtrlEnter :: Tb2Key

keyCtrlM :: Tb2Key

keyCtrlN :: Tb2Key

keyCtrlO :: Tb2Key

keyCtrlP :: Tb2Key

keyCtrlQ :: Tb2Key

keyCtrlR :: Tb2Key

keyCtrlS :: Tb2Key

keyCtrlT :: Tb2Key

keyCtrlU :: Tb2Key

keyCtrlV :: Tb2Key

keyCtrlW :: Tb2Key

keyCtrlX :: Tb2Key

keyCtrlY :: Tb2Key

keyCtrlZ :: Tb2Key

keyCtrlEsc :: Tb2Key

keyCtrlLsqBracket :: Tb2Key

keyCtrl3 :: Tb2Key

keyCtrl4 :: Tb2Key

keyCtrlBackslash :: Tb2Key

keyCtrl5 :: Tb2Key

keyCtrlRsqBracket :: Tb2Key

keyCtrl6 :: Tb2Key

keyCtrl7 :: Tb2Key

keyCtrlSlash :: Tb2Key

keyCtrlUnderscore :: Tb2Key

keySpace :: Tb2Key

keyBackspace2 :: Tb2Key

keyCtrl8 :: Tb2Key

keyF1 :: Tb2Key

keyF2 :: Tb2Key

keyF3 :: Tb2Key

keyF4 :: Tb2Key

keyF5 :: Tb2Key

keyF6 :: Tb2Key

keyF7 :: Tb2Key

keyF8 :: Tb2Key

keyF9 :: Tb2Key

keyF10 :: Tb2Key

keyF11 :: Tb2Key

keyF12 :: Tb2Key

keyInsert :: Tb2Key

keyDelete :: Tb2Key

keyHome :: Tb2Key

keyEnd :: Tb2Key

keyPgUp :: Tb2Key

keyPgDn :: Tb2Key

keyArrowUp :: Tb2Key

keyArrowDown :: Tb2Key

keyArrowLeft :: Tb2Key

keyArrowRight :: Tb2Key

keyBackTab :: Tb2Key

keyMouseLeft :: Tb2Key

keyMouseRight :: Tb2Key

keyMouseMiddle :: Tb2Key

keyMouseRelease :: Tb2Key

keyMouseWheelUp :: Tb2Key

keyMouseWheelDown :: Tb2Key

Modifiers

newtype Tb2Mod

Constructors

Tb2Mod CInt

modAlt :: Tb2Mod

modCtrl :: Tb2Mod

modShift :: Tb2Mod

modMotion :: Tb2Mod

Caps

newtype Tb2Cap

Constructors

Tb2Cap CInt

capF1 :: Tb2Cap

capF2 :: Tb2Cap

capF3 :: Tb2Cap

capF4 :: Tb2Cap

capF5 :: Tb2Cap

capF6 :: Tb2Cap

capF7 :: Tb2Cap

capF8 :: Tb2Cap

capF9 :: Tb2Cap

capF10 :: Tb2Cap

capF11 :: Tb2Cap

capF12 :: Tb2Cap

capInsert :: Tb2Cap

capDelete :: Tb2Cap

capHome :: Tb2Cap

capEnd :: Tb2Cap

capPgUp :: Tb2Cap

capPgDn :: Tb2Cap

capArrowUp :: Tb2Cap

capArrowDown :: Tb2Cap

capArrowLeft :: Tb2Cap

capArrowRight :: Tb2Cap

capBackTab :: Tb2Cap

capEnterCA :: Tb2Cap

capExitCA :: Tb2Cap

capShowCursor :: Tb2Cap

capHideCursor :: Tb2Cap

capClearScreen :: Tb2Cap

capSGR0 :: Tb2Cap

capUnderline :: Tb2Cap

capBold :: Tb2Cap

capBlink :: Tb2Cap

capItalic :: Tb2Cap

capReverse :: Tb2Cap

capEnterKeypad :: Tb2Cap

capExitKeypad :: Tb2Cap

Event types

newtype Tb2EventType

Constructors

Tb2EventType CInt

eventKey :: Tb2EventType

eventResize :: Tb2EventType

eventMouse :: Tb2EventType

Colors & Attributes

newtype Tb2ColorAttr

Constructors

Tb2ColorAttr CInt

colorBlack :: Tb2ColorAttr

colorRed :: Tb2ColorAttr

colorGreen :: Tb2ColorAttr

colorYellow :: Tb2ColorAttr

colorBlue :: Tb2ColorAttr

colorMagenta :: Tb2ColorAttr

colorCyan :: Tb2ColorAttr

colorWhite :: Tb2ColorAttr

colorDefault :: Tb2ColorAttr

attrBold :: Tb2ColorAttr

attrUnderline :: Tb2ColorAttr

attrReverse :: Tb2ColorAttr

attrItalic :: Tb2ColorAttr

attrBlink :: Tb2ColorAttr

Input modes

newtype Tb2Input

Constructors

Tb2Input CInt

inputCurrent :: Tb2Input

inputEsc :: Tb2Input

inputAlt :: Tb2Input

inputMouse :: Tb2Input

Output modes

newtype Tb2Output

Constructors

Tb2Output CInt

outputCurrent :: Tb2Output

outputNormal :: Tb2Output

output256 :: Tb2Output

output216 :: Tb2Output

outputGrayscale :: Tb2Output